+++ tests/EnvTest.php
@@ -7,17 +7,27 @@ class EnvTest extends TestCase
 {
     protected function setUp(): void
     {
-        foreach ([
-            'APP_NAME',
-            'FEATURE_ENABLED',
-            'FEATURE_DISABLED',
-            'OPTIONAL_VALUE',
-            'EMPTY_VALUE',
-            'SSR_URL',
-            'QUOTED_SINGLE',
-            'QUOTED_DOUBLE'
-        ] as $key) {
+        foreach (
+            [
+                'APP_NAME',
+                'FEATURE_ENABLED',
+                'FEATURE_DISABLED',
+                'OPTIONAL_VALUE',
+                'EMPTY_VALUE',
+                'SSR_URL',
+                'QUOTED_SINGLE',
+                'QUOTED_DOUBLE',
+                'QUOTED_WITH_COMMENT',
+                'WITH_SPACES',
+                'INLINE_COMMENT',
+                'ESCAPED_HASH',
+                'ESCAPED_EQUALS',
+                'MULTILINE',
+                'EXPORTED_VALUE'
+            ] as $key
+        ) {
             unset($_ENV[$key], $_SERVER[$key]);
+
             putenv($key);
         }
 
@@ -53,7 +63,7 @@ class EnvTest extends TestCase
     public function testEnvHandlesSingleAndDoubleQuotes(): void
     {
         $this->assertSame('Hello Single', Env::get('QUOTED_SINGLE'));
-        $this->assertSame('Hello Double', Env::get('QUOTED_DOUBLE'));
+        $this->assertSame('Hello "Double"', Env::get('QUOTED_DOUBLE'));
     }
 
     public function testEnvReturnsDefaultIfNotSet(): void
@@ -68,5 +78,42 @@ class EnvTest extends TestCase
         Env::load(__DIR__ . '/.env.example');
 
         $this->assertSame('ManualApp', Env::get('APP_NAME'));
+    }
+
+    public function testEnvHandlesInlineComments(): void
+    {
+        $this->assertSame('works', Env::get('INLINE_COMMENT'));
+        $this->assertSame('Hello # not comment', Env::get('QUOTED_WITH_COMMENT'));
+    }
+
+    public function testEnvHandlesSpacesAroundEquals(): void
+    {
+        $this->assertSame('spaced value', Env::get('WITH_SPACES'));
+    }
+
+    public function testEnvHandlesEscapedChars(): void
+    {
+        $this->assertSame('foo#bar', Env::get('ESCAPED_HASH'));
+        $this->assertSame('foo=bar', Env::get('ESCAPED_EQUALS'));
+    }
+
+    public function testEnvHandlesEscapedNewlines(): void
+    {
+        $this->assertSame("line1\nline2\rline3", Env::get('MULTILINE'));
+    }
+
+    public function testEnvHandlesExportPrefix(): void
+    {
+        $this->assertSame('exported', Env::get('EXPORTED_VALUE'));
+    }
+
+    public function testEnvDoesNotOverrideGetenv(): void
+    {
+        unset($_ENV['APP_NAME'], $_SERVER['APP_NAME']);
+        putenv('APP_NAME=SystemApp');
+
+        Env::load(__DIR__ . '/.env.example');
+
+        $this->assertSame('SystemApp', Env::get('APP_NAME'));
     }
 }